foxglove_sdk/ethernet_ip_integration/requirements.txt
foxglove-sdk
cpppo
pylogix
numpy
